refactor : pages components 분리
