+++ week4/find-your-github/src/components/SearchResult.tsx
